requirements.txt
# JAX ecosystem for DIAL-MPC controller
jax[cpu]>=0.4.0
# For CUDA GPU support, replace above with: jax[cuda11_pip]>=0.4.0
flax>=0.7.0
optax>=0.1.0
chex>=0.1.0

# Scientific computing
numpy>=1.21.0
scipy>=1.9.0

# Visualization
matplotlib>=3.5.0

# Distributed system dependencies
asyncio-mqtt>=0.13.0
websockets>=11.0.0
pyzmq>=25.0.0
tornado>=6.1  # For ZeroMQ asyncio compatibility on Windows
aiofiles>=23.0.0

# Configuration management
pyyaml>=6.0

# Performance monitoring and logging
tqdm>=4.64.0
tensorboard>=2.10.0
psutil>=5.9.0

# Data processing and analysis
pandas>=1.5.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Web demo dependencies
flask>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.33.0

# Optional: For neural scene oracles
torch>=2.0.0  # Uncomment if using PyTorch-based neural fields
torchvision>=0.15.0  # Uncomment if using PyTorch-based neural fields
torchaudio

# Optional: For Isaac Sim integration
# omni-isaac-core  # Install through Isaac Sim package manager

# Optional: For cloud deployment
# google-cloud-storage>=2.10.0  # Uncomment for GCP integration
# boto3>=1.26.0  # Uncomment for AWS integration

# RL requirements
gymnasium
stable-baselines3[extra]

# Added from the code block
pymavlink
pyserial
